remove navigationservice and isAppready flag
