+++ ios-app/UI/MainViewController.swift
@@ -149,6 +149,7 @@ class MainViewController: UIViewController {
         }
         
         window.rootViewController = viewController
+        DeepLinkRouter.shared.flushPendingDeepLink()
     }
     
     private var navigationBarHeight: CGFloat {
